pyproject.toml
[project]
name = "purple-builder-agent"
version = "1.0.0"
description = "A high-performance purple agent for the build_what_i_mean benchmark - AgentBeats Phase 2"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.34.0",
    "httpx>=0.28.0",
    "anthropic>=0.52.0",
    "openai>=1.70.0",
    "litellm>=1.60.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
